requirements.txt

fastapi
uvicorn
openai
pytesseract
pdfplumber
python-docx
pandas
Pillow
faiss-cpu
sentence-transformers
python-multipart
python-dotenv
streamlit
